+++ src/provenancelib.c
@@ -490,3 +490,33 @@ declare_set_secctx_fcn(provenance_secctx_track, PROV_SEC_TRACKED);
 declare_set_secctx_fcn(provenance_secctx_propagate, PROV_SEC_TRACKED|PROV_SEC_PROPAGATE);
 declare_set_secctx_fcn(provenance_secctx_delete, PROV_SEC_DELETE);
 declare_get_secctx_fcn(provenance_secctx);
+
+#define declare_set_cgroup_fcn(fcn_name, operation) int fcn_name (const uint32_t cid){\
+  struct cgroupinfo filter;\
+  int rc;\
+  int fd = open(PROV_CGROUP_FILTER, O_WRONLY);\
+  if( fd < 0 ){\
+    return fd;\
+  }\
+  filter.cid = cid;\
+  filter.op = operation;\
+  rc = write(fd, &filter, sizeof(struct cgroupinfo));\
+  close(fd);\
+  return rc;\
+}
+
+#define declare_get_cgroup_fcn(fcn_name) int fcn_name ( struct cgroupinfo* filters, size_t length ){\
+  int rc;\
+  int fd = open(PROV_CGROUP_FILTER, O_RDONLY);\
+  if( fd < 0 ){\
+    return fd;\
+  }\
+  rc = read(fd, filters, length);\
+  close(fd);\
+  return rc;\
+}
+
+declare_set_cgroup_fcn(provenance_cgroup_track, PROV_CGROUP_TRACKED);
+declare_set_cgroup_fcn(provenance_cgroup_propagate, PROV_CGROUP_TRACKED|PROV_CGROUP_PROPAGATE);
+declare_set_cgroup_fcn(provenance_cgroup_delete, PROV_CGROUP_DELETE);
+declare_get_cgroup_fcn(provenance_cgroup);