+++ tool/tool.c
@@ -46,6 +46,7 @@
 #define ARG_PROPAGATE_FILTER_EDGE       "--edge-propagate-filter"
 #define ARG_FILTER_RESET                "--reset-filter"
 #define ARG_SECCTX_FILTER               "--track-secctx"
+#define ARG_CGROUP_FILTER               "--track-cgroup"
 
 #define ANSI_COLOR_RED     "\x1b[31m"
 #define ANSI_COLOR_GREEN   "\x1b[32m"
@@ -75,6 +76,7 @@ void usage( void ){
   printf(CMD_COLORED CMD_PARAMETER("ip/mask:port") CMD_PARAMETER("track/propagate/delete") " track/propagate on bind.\n", ARG_TRACK_IPV4_INGRESS);
   printf(CMD_COLORED CMD_PARAMETER("ip/mask:port") CMD_PARAMETER("track/propagate/delete") " track/propagate on connect.\n", ARG_TRACK_IPV4_EGRESS);
   printf(CMD_COLORED CMD_PARAMETER("security context") CMD_PARAMETER("track/propagate/delete") " track/propagate based on security context.\n", ARG_SECCTX_FILTER);
+  printf(CMD_COLORED CMD_PARAMETER("cgroup no") CMD_PARAMETER("track/propagate/delete") " track/propagate based on cgroup.\n", ARG_CGROUP_FILTER);
   printf(CMD_COLORED CMD_PARAMETER("type") CMD_PARAMETER("bool") " set node filter.\n", ARG_FILTER_NODE);
   printf(CMD_COLORED CMD_PARAMETER("type") CMD_PARAMETER("bool") " set edge filter.\n", ARG_FILTER_EDGE);
   printf(CMD_COLORED CMD_PARAMETER("type") CMD_PARAMETER("bool") " set propagate node filter.\n", ARG_PROPAGATE_FILTER_NODE);
@@ -112,6 +114,7 @@ void state( void ){
   uint64_t filter=0;
   struct prov_ipv4_filter filters[100];
   struct secinfo sec_filters[100];
+  struct cgroupinfo cgroup_filters[100];
   int size;
   uint32_t machine_id;
   int i;
@@ -176,10 +179,22 @@ void state( void ){
   size = provenance_secctx(sec_filters, 100*sizeof(struct secinfo));
   printf("Security context filter (%ld).\n", size/sizeof(struct secinfo));
   for(i = 0; i < size/sizeof(struct secinfo); i++){
-    printf("%s", sec_filters[i].secctx);
-    if((filters[i].op&PROV_SEC_PROPAGATE) == PROV_SEC_PROPAGATE){
+    printf("%s ", sec_filters[i].secctx);
+    if((sec_filters[i].op&PROV_SEC_PROPAGATE) == PROV_SEC_PROPAGATE){
       printf("propagate");
-    }else if((filters[i].op&PROV_SEC_TRACKED) == PROV_SEC_TRACKED){
+    }else if((sec_filters[i].op&PROV_SEC_TRACKED) == PROV_SEC_TRACKED){
+      printf("track");
+    }
+    printf("\n");
+  }
+
+  size = provenance_cgroup(cgroup_filters, 100*sizeof(struct cgroupinfo));
+  printf("CGroup filter (%ld).\n", size/sizeof(struct cgroupinfo));
+  for(i = 0; i < size/sizeof(struct cgroupinfo); i++){
+    printf("%u ", cgroup_filters[i].cid);
+    if((cgroup_filters[i].op&PROV_CGROUP_PROPAGATE) == PROV_CGROUP_PROPAGATE){
+      printf("propagate");
+    }else if((cgroup_filters[i].op&PROV_CGROUP_TRACKED) == PROV_CGROUP_TRACKED){
       printf("track");
     }
     printf("\n");
@@ -414,6 +429,21 @@ int main(int argc, char *argv[]){
 
     if(err < 0){
       perror("Could not change security context filter.\n");
+    }
+    return 0;
+  }
+  MATCH_ARGS(argv[1], ARG_CGROUP_FILTER){
+    CHECK_ATTR_NB(argc, 4);
+    if( is_str_propagate( argv[3]) ){
+      err = provenance_cgroup_propagate(strtoul(argv[2], NULL, 0));
+    }else if( is_str_track(argv[3])){
+      err = provenance_cgroup_track(strtoul(argv[2], NULL, 0));
+    }else if( is_str_delete(argv[3])){
+      err = provenance_cgroup_delete(strtoul(argv[2], NULL, 0));
+    }
+
+    if(err < 0){
+      perror("Could not change CGroup filter.\n");
     }
     return 0;
   }